% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhw14.R
\name{fhw14}
\alias{fhw14}
\title{fhw15}
\usage{
fhw14(x, p)
}
\arguments{
\item{x}{array of data}

\item{p}{array of probability}
}
\value{
mean, variance of input data
}
\description{
Calculate the mu and variance with data weighted with probability sums to 1; without GIEMO
}
