% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhw25.R
\name{fhw25}
\alias{fhw25}
\title{fhw25}
\usage{
fhw25(X, MARGIN, FUN, ...)
}
\arguments{
\item{X}{is a matrix of any type a matrix can have (numeric, character, logical, complex)}

\item{MARGIN}{is either (the number) 1 or (the number)}

\item{FUN}{is an R function that maps vectors to vectors}

\item{...}{is passed to FUN, that is, any arguments to myapply that do not match X, MARGIN, or
FUN are passed to FUN whenever it is called by fhw25}
}
\value{
result of the matrix calculation using FUN
}
\description{
a function just like the function array in the R base package, but is simpler.
}
