% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhw26.R
\name{fhw26}
\alias{fhw26}
\title{fquiz22}
\usage{
fhw26(x)
}
\arguments{
\item{x}{3d matrix}
}
\value{
median to the three two-dimensional margins (indexed by pairs of indices) 
and the three one-dimensional margins (indexed by single indices) of this array.
}
\description{
Loading a 3d matrix; 
Apply the R function median to the three two-dimensional margins (indexed by pairs of indices) 
and the three one-dimensional margins (indexed by single indices) of this array.
}
