% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fquiz11.R
\name{fquiz11}
\alias{fquiz11}
\title{fquiz11}
\usage{
fquiz11(x)
}
\arguments{
\item{x}{Name of your data}
}
\value{
mean, variance, standard deviation of input data
}
\description{
Calculate the mean, variance, standard deviation of input data
}
