% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fquiz12.R
\name{fquiz12}
\alias{fquiz12}
\title{fquiz12}
\usage{
fquiz12(x)
}
\arguments{
\item{x}{Name of your data}
}
\value{
mean, variance, standard deviation of input data
}
\description{
Return mean, variance, standard deviation of input data, with GIEMO
}
