% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fquiz13.R
\name{fquiz13}
\alias{fquiz13}
\title{fquiz13}
\usage{
fquiz13(x)
}
\arguments{
\item{x}{Name of your data}
}
\value{
maximum likelihood estimate of input data Using dgmma
}
\description{
Using dgamma calculte maximum likelihood estimate of input data
}
