% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fquiz21.R
\name{fquiz21}
\alias{fquiz21}
\title{fquiz21}
\usage{
fquiz21(a, x)
}
\arguments{
\item{a}{numeric matrix A}

\item{x}{numeric vector x}
}
\value{
value of x^T A^{-1} x
}
\description{
Give a numeric matrix A and a numeric vector x, calculates x^T A^{-1} x;with GIEMO
}
