% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optfun.R
\name{ga}
\alias{ga}
\title{optfun
 @examples
optfun(x$x1,ga,c(0,3))
optfun(x$x1,cau,c(-10,10))
optfun(x$x1,bin,c(-100,0))}
\usage{
ga(theta, x)
}
\description{
optfun
 @examples
optfun(x$x1,ga,c(0,3))
optfun(x$x1,cau,c(-10,10))
optfun(x$x1,bin,c(-100,0))
}
