% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fquiz22.R
\name{\%fquiz22\%}
\alias{\%fquiz22\%}
\title{fquiz22}
\usage{
a \%fquiz22\% x
}
\arguments{
\item{a}{numeric matrix}

\item{x}{numeric vector}
}
\value{
value of  x^T A^{-1} x
}
\description{
fquiz22 invoked as: a %function% x; a  is numeric matrix  and x is numeric vector. fquiz22 calculates x^T A^{-1} x; With GIEMO
}
