% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hggplot.R
\name{hggplot}
\alias{hggplot}
\title{hggplot}
\usage{
hggplot(...)
}
\arguments{
\item{data}{Name of your data}

\item{mapping}{Default list of aesthetic mappings to use for plot. If not specified, must be suppled in each layer added to the plot.}

\item{environment}{an variable defined in the aesthetic mapping is not found in the data, this function will look for it in this environment.}
}
\value{
Scatterplot of you data with your specified x and y and other value.
}
\description{
Scatterplot your data, using the same syntax as ggplot2::ggplot()
}
