% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{x}
\alias{x}
\title{This is data to be included in my package}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 200 rows and 3 columns.}
\usage{
x
}
\description{
This is data to be included in my package
}
\references{
\url{http://www.stat.umn.edu/geyer/3701/data/p4p3.csv}
}
\keyword{datasets}
